Move tests to the testkit
