+++ exonum/src/runtime/configuration/transactions.rs
@@ -105,7 +105,7 @@ pub enum ConfigurationTransactions {
 
 impl ConfigurationTransactions {
     #[doc(hidden)]
-    #[cfg(test)]
+    // TODO: pub only for testing.
     pub fn from_raw(message: Signed<RawTransaction>) -> ConfigurationTransactions {
         use crate::blockchain::TransactionSet;
         use std::ops::Deref;
@@ -304,7 +304,9 @@ impl Transaction for Propose {
     }
 }
 
-struct VotingContext {
+// TODO: Public only for testing.
+#[derive(Debug)]
+pub struct VotingContext {
     decision: VotingDecision,
     author: PublicKey,
     cfg_hash: Hash,
@@ -312,7 +314,7 @@ struct VotingContext {
 
 impl VotingContext {
     /// Creates new `VotingContext` from `VotingDecision` and author key.
-    fn new(decision: VotingDecision, author: PublicKey, cfg_hash: Hash) -> Self {
+    pub fn new(decision: VotingDecision, author: PublicKey, cfg_hash: Hash) -> Self {
         VotingContext {
             author,
             decision,
@@ -325,7 +327,7 @@ impl VotingContext {
     /// # Return value
     ///
     /// Returns a configuration this transaction is for on success, or an error (if any).
-    fn precheck(&self, snapshot: &dyn Snapshot) -> Result<StoredConfiguration, ServiceError> {
+    pub fn precheck(&self, snapshot: &dyn Snapshot) -> Result<StoredConfiguration, ServiceError> {
         use self::ServiceError::*;
 
         let following_config = CoreSchema::new(snapshot).following_configuration();
@@ -445,49 +447,5 @@ impl Transaction for VoteAgainst {
         );
 
         Ok(())
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use exonum_testkit::{TestKit, TestKitBuilder};
-
-    use super::{Hash, VotingContext};
-    use crate::{
-        config::ConfigurationServiceConfig,
-        errors::Error as ServiceError,
-        schema::VotingDecision,
-        tests::{new_tx_config_vote, new_tx_config_vote_against},
-        Service as ConfigurationService,
-    };
-
-    #[test]
-    fn test_vote_without_propose() {
-        let testkit: TestKit = TestKitBuilder::validator()
-            .with_validators(4)
-            .with_service(ConfigurationService {
-                config: ConfigurationServiceConfig::default(),
-            })
-            .create();
-
-        let hash = Hash::default();
-
-        let illegal_vote = new_tx_config_vote(&testkit.network().validators()[3], hash);
-
-        let decision = VotingDecision::Yea(illegal_vote.hash());
-        let author = illegal_vote.author();
-        let vote = VotingContext::new(decision, author, hash);
-
-        let vote_result = vote.precheck(testkit.snapshot().as_ref());
-
-        let illegal_vote_against =
-            new_tx_config_vote_against(&testkit.network().validators()[3], hash);
-        let decision = VotingDecision::Yea(illegal_vote_against.hash());
-        let author = illegal_vote_against.author();
-        let vote_against = VotingContext::new(decision, author, hash);
-        let vote_against_result = vote_against.precheck(testkit.snapshot().as_ref());
-
-        assert_matches!(vote_result, Err(ServiceError::UnknownConfigRef(_)));
-        assert_matches!(vote_against_result, Err(ServiceError::UnknownConfigRef(_)));
     }
 }
